Reorganise code and add paged github API calls
* Move the PR fetcher to a separate class
* Move the output formatter to a separate class
* Repeat the github API call until no more data

The cloud platform team has >100 github repos, but
the API call was only fetching open PRs for the
first 100 repositories.

This change uses the returned 'pageInfo' data to
repeat the API call until there is no more data to
retrieve.
