+++ lib/pr.rb
@@ -1,0 +1,9 @@
+require 'bundler/setup'
+require "graphql/client"
+require "graphql/client/http"
+require 'date'
+
+require "#{File.dirname(__FILE__)}/pr/fetcher"
+require "#{File.dirname(__FILE__)}/pr/formatter"
+
+class PR; end
